LPS-74897 Squash InitialThreadLocal/AutoResetThreadLocal into CentralizedThreadLocal
