--- modules/apps/foundation/petra/petra-lang/src/main/java/com/liferay/petra/lang/InitialThreadLocal.java
@@ -1,61 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.petra.lang;
-
-import java.util.function.Supplier;
-
-/**
- * @author Brian Wing Shun Chan
- * @author Shuyang Zhou
- */
-public class InitialThreadLocal<T> extends CentralizedThreadLocal<T> {
-
-	public InitialThreadLocal(String name, Supplier<T> supplier) {
-		this(name, supplier, false);
-	}
-
-	public InitialThreadLocal(
-		String name, Supplier<T> supplier, boolean shortLived) {
-
-		super(shortLived);
-
-		_name = name;
-
-		if (supplier == null) {
-			_supplier = () -> null;
-		}
-		else {
-			_supplier = supplier;
-		}
-	}
-
-	@Override
-	public String toString() {
-		if (_name == null) {
-			return super.toString();
-		}
-
-		return _name;
-	}
-
-	@Override
-	protected T initialValue() {
-		return _supplier.get();
-	}
-
-	private final String _name;
-	private final Supplier<T> _supplier;
-
-}
